docker: Cleanup builder caches on every run
Since we're using DOCKER_BUILDKIT and it has some issues in handling
caches for pulled images, we need to make sure that builder caches
are not causing problems.

It's fairly safe to prune builder caches as it does not affect already
running builds and it does not remove images/containers.